% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billmillr-package.R
\docType{package}
\name{billmillr-package}
\alias{billmillr}
\alias{billmillr-package}
\title{billmillr: Rethinking the Bill Miller Problem}
\description{
This package calculates the likelihood that any one trader out of n traders will beat the market in at least a k-years sequence out of a t-years time period by considering independent and definite start and stop periods of such a k-years sequence given a t-years time period for the nth trader.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mghoff/billmillr}
  \item \url{https://mghoff.github.io/billmillr/}
}

}
\author{
\strong{Maintainer}: Matthew Hoff \email{matthew.g.hoff@gmail.com}

}
\keyword{internal}
