% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{odds_of_streak}
\alias{odds_of_streak}
\title{Odds of a single streak}
\usage{
odds_of_streak(numCoins = 10, minHeads = 5, probHeads = 0.5, .saved = NULL)
}
\arguments{
\item{numCoins}{integer, total number of coin flips}

\item{minHeads}{integer, minimum number of heads to to obtained in a single streak}

\item{probHeads}{number, probability of obtaining a heads on a single coin toss}

\item{.saved}{data frame, contains previously computed probabilities for a given index;
used to speed up recursion so not to recalculate every probability should an already seen
scenario of numCoins, minHeads, and probHeads arises.}
}
\value{
numeric, odds of streak
}
\description{
Calculate the odds of a single streak of at least k heads out of n coin tosses given
the probability of heads (tails) is p (q = 1-p).
}
\examples{
odds_of_streak(numCoins = 10, minHeads = 3, probHeads = 0.5)
}
