% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{prob_of_at_least_k}
\alias{prob_of_at_least_k}
\title{Odds of at least K streaks}
\usage{
prob_of_at_least_k(N, K, p)
}
\arguments{
\item{N}{integer, total number of coin tossers.}

\item{K}{integer, number of streaks.}

\item{p}{number, value between 0 & 1 representing the probability of a single streak.}
}
\value{
a number between 0 and 1
}
\description{
Calculate the probability of the occurrence of at least K-streaks out of N-events
given the odds of stated streak.
}
\examples{
prob_of_at_least_k(
 N = 8,
 K = 1,
 p = odds_of_streak(num_coins = 5, min_heads = 3, prob_heads = 0.5)
)
}
