% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billmillr.R
\name{probOfAtLeastK}
\alias{probOfAtLeastK}
\title{Odds of at least K streaks}
\usage{
probOfAtLeastK(N, K, P)
}
\arguments{
\item{N}{integer, total number of coin tossers}

\item{K}{integer, number of streaks}

\item{P}{number, probability of a single streak}
}
\value{
a number
}
\description{
Calculate the probability of the occurrence of at least K-streaks out of N-events
given the odds of stated streak.
}
\examples{
probOfAtLeastK(N = 100, K = 1, P = oddsOfStreak())
}
