% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulr.R
\name{count_runs}
\alias{count_runs}
\title{Count Runs}
\usage{
count_runs(trials = 1000, flips = 40, streak_length = 15)
}
\arguments{
\item{trials}{number, number of attempts to flip a coin n times}

\item{flips}{number, number of sequential coin flips in which a streak could be counted}

\item{streak_length}{number, length of streak of interest}
}
\value{
number
}
\description{
Out of a given number of trials of flipping a coin n times, count the number of
times a run/streak of length k occurs
}
\examples{
count_runs()
}
