% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulr.R
\name{find_longest_run}
\alias{find_longest_run}
\title{Find the Longest Run}
\usage{
find_longest_run(binom_sample)
}
\arguments{
\item{binom_sample, }{numeric vector - a binomial sample; e.g. \code{sample(c(0, 1), size = 5, replace = TRUE)}}
}
\value{
number, representing the longest streak (run of "1") in the provided sample
}
\description{
Finds the single longest run ("streak") of the value "1"
provided in a binomial sample of size n.
}
\examples{
find_longest_run(binom_sample = sample(c(0, 1), size = 25, replace = TRUE))
}
