% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{prob_of_at_least_k}
\alias{prob_of_at_least_k}
\title{Odds of at least K streaks}
\usage{
prob_of_at_least_k(N, K, P)
}
\arguments{
\item{N}{integer, total number of coin tossers}

\item{K}{integer, number of streaks}

\item{P}{number, probability of a single streak}
}
\value{
a number
}
\description{
Calculate the probability of the occurrence of at least K-streaks out of N-events
given the odds of stated streak.
}
\examples{
prob_of_at_least_k(N = 100, K = 1, P = odds_of_streak())
}
