% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{run_simulation}
\alias{run_simulation}
\title{Run Simulation}
\usage{
run_simulation(iters = 100, trials = 1000, flips = 40, streak_length = 15)
}
\arguments{
\item{iters}{number, number of simulation iterations to consider}

\item{trials}{number, see \code{count_runs()}}

\item{flips}{number, see \code{count_runs()}}

\item{streak_length}{number, see \code{count_runs()}}
}
\value{
data frame
}
\description{
Iterate over \code{count_runs()}, counting the number of applicable trials where
such a trial is defined as having the occurrence of a streak of at least the
designated length. Additionally, keep a running tally of the cumulative likelihood
of obtaining exactly zero streaks.
}
\details{
This function outputs a data frame with three columns:
1.) the iteration number,
2.) the count of applicable trials in that iteration, and
3.) the cumulative probability of obtaining zero streaks.
}
\examples{
tail(run_simulation())
}
